New panel and more info.
Added in the campaign panel, which holds text info for quick reference in
your campaign notes. Added in health and level areas to the results panel.
And made the whole attributes thing kosher. Going to go through the races
and classes to convert the rest of them... After I get the health
generation working.